refactor: move Tabs components into Grapher
