--- packages/@ourworldindata/components/src/Tabs/ExpandableTabs.scss
@@ -1,13 +1,0 @@
-.ExpandableTabs__button {
-    color: #5b5b5b;
-    border-color: #f2f2f2 !important; // !important overwrites style definitions in grapher.scss
-    background: #f2f2f2 !important;
-
-    &:hover {
-        background: #e7e7e7 !important;
-    }
-
-    svg {
-        margin-right: 6px;
-    }
-}
