--- packages/@ourworldindata/components/src/Tabs/Tabs.scss
@@ -1,47 +1,0 @@
-.Tabs {
-    --border: var(--tab-border, #{$blue-10});
-    --text: var(--tab-text, #858585);
-    --text-active: var(--text-text-active, #{$blue-90});
-    --background: var(--tab-background, #fff);
-    --background-active: var(--tab-background-active, #{$accent-pale-blue});
-
-    margin: 16px 0;
-
-    &--horizontal-scroll {
-        overflow-x: auto;
-        white-space: nowrap;
-        -webkit-overflow-scrolling: touch;
-        scrollbar-width: none;
-    }
-
-    &::-webkit-scrollbar {
-        display: none;
-    }
-
-    &__tab {
-        margin: 0 8px 8px 0;
-        padding: 8px 16px;
-        border: 1px solid var(--border) !important; // !important overwrites style definitions in grapher.scss
-        background: var(--background) !important;
-        color: var(--text);
-        font-size: 0.8125rem;
-        font-weight: 500;
-        letter-spacing: 0.01em;
-        white-space: nowrap;
-        max-width: 240px;
-        text-overflow: ellipsis;
-        overflow: hidden;
-
-        &:hover {
-            cursor: pointer;
-            background-color: $gray-10 !important;
-            border-color: $gray-10 !important;
-        }
-
-        &[aria-selected="true"] {
-            background: var(--background-active) !important;
-            border-color: var(--background-active) !important;
-            color: var(--text-active);
-        }
-    }
-}
