+++ packages/@ourworldindata/components/src/index.ts
@@ -17,8 +17,6 @@ export {
 } from "./GdocsUtils.js"
 
 export { ExpandableToggle } from "./ExpandableToggle/ExpandableToggle.js"
-export { Tabs } from "./Tabs/Tabs.js"
-export { ExpandableTabs } from "./Tabs/ExpandableTabs.js"
 
 export {
     makeSource,
