+++ packages/@ourworldindata/grapher/src/core/grapher.scss
@@ -6,8 +6,6 @@
 
 @import "../../../components/src/CodeSnippet/code-snippet.scss";
 @import "../../../components/src/ExpandableToggle/ExpandableToggle.scss";
-@import "../../../components/src/Tabs/Tabs.scss";
-@import "../../../components/src/Tabs/ExpandableTabs.scss";
 
 @import "../../../components/src/IndicatorSources/IndicatorSources.scss";
 @import "../../../components/src/IndicatorProcessing/IndicatorProcessing.scss";
@@ -73,6 +71,8 @@ $zindex-controls-drawer: 140;
     @import "../header/Header.scss";
     @import "../modal/SourcesKeyDataTable.scss";
     @import "../modal/SourcesDescriptions.scss";
+    @import "../tabs/Tabs.scss";
+    @import "../tabs/ExpandableTabs.scss";
 }
 
 // These rules are currently used elsewhere in the site. e.g. Explorers
