+++ packages/@ourworldindata/grapher/src/tabs/ExpandableTabs.scss
@@ -1,0 +1,13 @@
+.ExpandableTabs__button {
+    color: $dark-text;
+    border-color: #f2f2f2;
+    background: #f2f2f2;
+
+    &:hover {
+        background: #e7e7e7;
+    }
+
+    svg {
+        margin-right: 6px;
+    }
+}
